+++ otabek/src/components/MovieCard.js
