+++ otabek/src/index.css
@@ -1,3 +1,7 @@
+@import url("https://fonts.googleapis.com/css2?family=Train+One&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Oxygen:wght@300;400;700&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Josefin+Sans:wght@300;400&display=swap");
+
 html {
   font-size: 10px;
 }
@@ -9,8 +13,18 @@ html {
 body {
   margin: 0;
   padding: 0;
-  background-color: rgb(244, 244, 244);
   color: #333;
+  background: #be93c5; /* fallback for old browsers */
+  background: -webkit-linear-gradient(
+    to right,
+    #7bc6cc,
+    #be93c5
+  ); /* Chrome 10-25, Safari 5.1-6 */
+  background: linear-gradient(
+    to right,
+    #7bc6cc,
+    #be93c5
+  ); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
 }
 
 p {
@@ -18,7 +32,8 @@ p {
 }
 
 small {
-  font-size: 1.2rem;
+  font-size: 1.6rem;
+  font-weight: 700;
 }
 
 .container {
@@ -30,6 +45,8 @@ small {
 .title {
   font-size: 4.4rem;
   text-align: center;
+  font-family: "Train One", cursive;
+  color: aliceblue;
 }
 
 .form {
@@ -37,18 +54,27 @@ small {
 }
 
 .label {
+  color: aliceblue;
   font-size: 1.2rem;
   margin-bottom: 0.2rem;
   text-transform: uppercase;
+  font-family: "Oxygen", sans-serif;
 }
 
 .input {
   font-size: 1.6rem;
+  font-family: "Oxygen", sans-serif;
   padding: 0.5rem 2rem;
   line-height: 2.8rem;
   border-radius: 20px;
   border: 1px solid #ddd;
   margin-bottom: 1rem;
+  outline: none;
+  background-color: rgb(229, 226, 226);
+}
+.input:hover {
+  line-height: 3rem;
+  background-color: aliceblue;
 }
 
 .button {
@@ -60,6 +86,8 @@ small {
   font-size: 1.4rem;
   cursor: pointer;
   transition: background-color 250ms;
+  outline: none;
+  font-family: "Oxygen", sans-serif;
 }
 
 .button:hover {
@@ -88,11 +116,15 @@ small {
   box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.25);
   margin-bottom: 2rem;
   background-color: white;
+  font-family: "Josefin Sans", sans-serif;
+  font-weight: 400;
 }
 
 .card--title {
   margin-bottom: 0.5rem;
   font-size: 3.2rem;
+  font-weight: 700;
+  font-family: "Oxygen", sans-serif;
 }
 
 .card--image {
